bug: fix content type to return html by default
Curl was responding with application/json as that is what I told it to
do. For some webpages like github for example, that would change the
response from the html to a json object, which would not have a title
html tag, causing the pattern matching to always return `nil`.